Allow changing the LIST_ID of a task.
Just changing the LIST_ID of a task would confuse sync adapters, since
the task might already have a SYNC_ID of a different task list or even
account. To support that case we emulate a copy & delete operation when
we detect a change of the LIST_ID. The original task is moved over to
the new list, while clearing all sync adapter fields (SYNC_*) and a new
task having the deleted flag set is created in the old list.
Since the _ID of the actual task will be the same, the UI can show the
same task.

This commit also fixes resolution of ORIGINAL_INSTANCE_SYNC_ID and
ORIGINAL_INSTANCE_ID when an exception is added or the master task is
inserted after an exception.

We also dropped CursorTaskAdapter since we don't need it. Initially it
was used wrongly for deleted tasks.